refactor: テキスト類似性検索関連の関数をmodules/similarity.pyに記述
